add sample to TestWorker
